Extract ConvertOffsetToTime (win)
