+++ src/ui/windows/TogglDesktop/TogglDesktop/TogglDesktop.csproj
@@ -401,6 +401,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="utilities\StringExtensions.cs" />
+    <Compile Include="utilities\TimelineUtils.cs" />
     <Compile Include="utilities\Utils.cs" />
     <Compile Include="utilities\Win32.cs" />
     <Compile Include="ui\windows\FeedbackWindow.xaml.cs">
