+++ src/ui/windows/TogglDesktop/TogglDesktop/ui/Resources/TimelineConstants.cs
@@ -20,13 +20,5 @@ namespace TogglDesktop.Resources
             {2, 50},
             {3, 25}
         };
-
-        public static ulong ConvertOffsetToTime(double height, DateTime date, double hourHeight)
-        {
-            var hours = 1.0 * height / hourHeight;
-            var dateTime = date.AddHours(hours);
-            var unixTime = Toggl.UnixFromDateTime(dateTime);
-            return unixTime >= 0 ? (ulong)unixTime : 0;
-        }
     }
 }
